Fix deserialization order to match serialization
The previous commit, 86240a3, addressed the hardcoding of all LP
deserialization being mapped to the first LP type. This fix was a
positive step but still left a previously unaddressed issue:

The deserialization was not done in the same order as serialization.
The result of this was that in simulations with more than one KP, LPs
may end up with GIDs and RNGs different than what they had been given
at the start of the original simulation.

This bug could have gone unnoticed in simulations with only one LP type
as there would not be any runtime errors generated unless change in GID
or RNG detection was done at the model level.

This commit addresses the issue and makes the deserialization match the
serialiation process.