CLI Argument Adjustment and Addition
This commit adjusts one command line argument so that the variable
that it sets via command line is globally accessible. (a)

This commit also adds one command line argument that gives the developer
some control over how events are serialized. (b)

(a) Prior to this commit, there was an io_store was not a ubiquitous
globally accessible command line argument despite it being critical
to the core operation of RIO. Making it a globally accessible variable
allows model files beyond the main function to know whether they are in
a load or save mode.

(b) Prior to this commit, upon serialization of events that were
scheduled for after the end of the original simulaiton, the timestamps
are adjusted so that upon restart it will act as if it's a new event at
a time equal to its original timestamp - the original simulation's end time.

e.g. Upon Checkpointing, RIO captures 1 event that happened 20 timesteps
after the end of the simulation (end=1000). While the actual timestamp is
1020, RIO will store the event with a timestamp of 1020 - 1000 = 20. Upon
restarting, the saved event will be scheduled with ross at timestep 20.

This behavior is restrictive to model developers who may keep tw_stime
variables in LP state. An example would be an LP keeping track of the
last time that a messaged ocurred and not creating new messages until
at least some specific number of timesteps have passed. If the simulation
restarts with events happening "in the past" from the LP point of view,
it would interfere with the determinism of the simulaiton.

e.g. Imagine an LP that will only create new events if the time since
its last newly created event is at least 20 timesteps. If the end of the
simulation is 1000, and the LP last created an event at 990 but is
scheduled to receive an event at 1020. It never gets that 1020 event
becasue the simulation ended. But if the simulation was restarted that
same event would get delivered after all at timestep 20. If the original
simulation had kept going beyond 1000, that event in question would lead
to another event being created by the LP as the time since its last
created event was 30 timesteps. But in the restarted simulation, the LP
wouldn't create another event as it is suppressing its own event creation
until time 1010.